renaming docusing-commons,js as lib.js
