+++ src/appmixer/docusign/esignature/RequestSignature/RequestSignature.js
@@ -1,5 +1,5 @@
 'use strict';
-const commons = require('../../docusign-commons');
+const commons = require('../../lib');
 
 /**
  * Get an envelope.
