+++ src/appmixer/docusign/lib.js
@@ -1,28 +1,234 @@
-module.exports = {
+'use strict';
+const docusign = require('docusign-esign');
 
-    /**
-     * Normalizes multiselect field values to ensure consistent array format.
-     * This function handles various input formats (strings, arrays, comma-separated values)
-     * and converts them to a proper array format for API consumption.
-     *
-     * @param {string|Array|undefined} value - The value to normalize
-     * @returns {Array|undefined} - Normalized array or undefined if input is empty
-     */
-    normalizeMultiselect(value) {
+const getAccessToken = async (integrationKey, secretKey, code, context) => {
 
-        if (!value) return undefined;
+    const isProduction = context.config.production === true || context.config.production === 'true';
+    let authorizationHeader = integrationKey + ':' + secretKey;
+    let buff = Buffer.from(authorizationHeader, 'utf8');
+    authorizationHeader = 'Basic ' + buff.toString('base64');
+    authorizationHeader = {
+        'Content-Type': 'application/x-www-form-urlencoded',
+        Authorization: authorizationHeader
+    };
 
-        // If already an array, return as is
-        if (Array.isArray(value)) return value;
+    const baseUrl = isProduction ? 'https://account.docusign.com/oauth/token' : 'https://account-d.docusign.com/oauth/token';
+    let { data: tokenResponse } = await context.httpRequest({
+        url: baseUrl,
+        method: 'POST',
+        data: `grant_type=authorization_code&code=${code}`,
+        headers: authorizationHeader
+    });
 
-        // If string, split by comma and trim whitespace
-        if (typeof value === 'string') {
-            const result = value.split(',').map(item => item.trim()).filter(item => item.length > 0);
-            return result.length > 0 ? result : undefined;
+    return tokenResponse;
+};
+
+const refreshToken = async (integrationKey, secretKey, refreshToken, context) => {
+
+    const isProduction = context.config.production === true || context.config.production === 'true';
+    let authorizationHeader = integrationKey + ':' + secretKey;
+    let buff = Buffer.from(authorizationHeader, 'utf8');
+    authorizationHeader = 'Basic ' + buff.toString('base64');
+    authorizationHeader = {
+        'Content-Type': 'application/x-www-form-urlencoded',
+        Authorization: authorizationHeader
+    };
+
+    const baseUrl = isProduction ? 'https://account.docusign.com/oauth/token' : 'https://account-d.docusign.com/oauth/token';
+    let { data: tokenResponse } = await context.httpRequest({
+        url: baseUrl,
+        method: 'POST',
+        data: `grant_type=refresh_token&refresh_token=${encodeURIComponent(refreshToken)}`,
+        headers: authorizationHeader
+    });
+
+    return tokenResponse;
+};
+
+
+const getEnvelope = async (args, accessToken) => {
+
+    let dsApiClient = new docusign.ApiClient();
+    dsApiClient.setBasePath(args.basePath + '/restapi/');
+    dsApiClient.addDefaultHeader('Authorization', 'Bearer ' + accessToken);
+    let envelopesApi = new docusign.EnvelopesApi(dsApiClient);
+    return await envelopesApi.getEnvelope(
+        args.accountId,
+        args.envelopeId,
+        { include: args.include && args.include.join(',') }
+    );
+};
+
+/**
+ * Create and send envelope.
+ * @param {Object} args
+ * @param {String} args.basePath
+ * @param {String} args.accountId
+ * @param {Object} args.envelopeArgs
+ * @param {Object} args.docs
+ * @param {String} accessToken
+ * @param {String} webhookUrl
+ * @return {Promise}
+ */
+const requestSignature = async (args, accessToken, webhookUrl) => {
+
+    let dsApiClient = new docusign.ApiClient();
+    dsApiClient.setBasePath(args.basePath + '/restapi/');
+    dsApiClient.addDefaultHeader('Authorization', 'Bearer ' + accessToken);
+    let envelopesApi = new docusign.EnvelopesApi(dsApiClient);
+    let results = null;
+    // Step 1. Make the envelope request body
+    let envelopeOptions = await makeEnvDefinition(args.envelopeArgs, args.docs, webhookUrl);
+    // Step 2. call Envelopes::create API method
+    // Exceptions will be caught by the calling function
+    results = await envelopesApi.createEnvelope(args.accountId, envelopeOptions);
+    return results;
+};
+
+/**
+ * Create envelope definition.
+ * @param {Object} args
+ * @param {Object} docs
+ * @param {String} webhookUrl
+ * @return {Promise}
+ */
+const makeEnvDefinition = async (args, docs, webhookUrl) => {
+
+    const envObject = {
+        documents:
+            docs.map((doc, idx) => {
+                return {
+                    documentBase64: doc.doc.toString('base64'),
+                    documentId: `${idx + 1}`,
+                    fileExtension: doc.fileExtension,
+                    name: doc.name || doc.fileName
+                };
+            }),
+        emailSubject: args.subject,
+        emailBlurb: args.message,
+        recipients: {
+            signers:
+                args.signers.AND.map((user, idx) => {
+                    return {
+                        email: user.email,
+                        name: user.name,
+                        recipientId: `${idx + 1}`
+                    };
+                })
+        },
+        status: 'sent',
+        eventNotification: {
+            useSoapInterface: false,
+            URL: webhookUrl,
+            loggingEnabled: 'true',
+            requireAcknowledgment: 'true',
+            eventData: {
+                version: 'restv2.1',
+                format: 'json',
+                includeData: [
+                    'tabs'
+                ]
+            },
+            envelopeEvents: [
+                {
+                    envelopeEventStatusCode: 'completed'
+                }
+            ],
+            recipientEvents: []
         }
+    };
+    return {
+        cdseMode: '',
+        changeRoutingOrder: 'true',
+        completedDocumentsOnly: 'false',
+        mergeRolesOnDraft: '',
+        tabLabelExactMatches: '',
+        envelopeDefinition: envObject
+    };
+};
 
-        // For other types, convert to string first then split
-        const result = String(value).split(',').map(item => item.trim()).filter(item => item.length > 0);
+/**
+ * Register of specific events to watch.
+ * @param {Object} args
+ * @param {String} args.basePath
+ * @param {Object} args.events
+ * @param {String} args.accountId
+ * @param {String} accessToken
+ * @param {String} webhookUrl
+ * @return {Promise}
+ */
+const registerDocusignWebhook = async (args, accessToken, webhookUrl) => {
+
+    let dsApiClient = new docusign.ApiClient();
+    dsApiClient.setBasePath(args.basePath + '/restapi/');
+    dsApiClient.addDefaultHeader('Authorization', 'Bearer ' + accessToken);
+    let connectApi = new docusign.ConnectApi(dsApiClient);
+    let connectCustomConfiguration = {
+        configurationType: 'custom',
+        urlToPublishTo: webhookUrl,
+        allUsers: 'true',
+        name: `Appmixer-${args.events.join('|')}`,
+        deliveryMode: 'SIM',
+        allowEnvelopePublish: 'true',
+        enableLog: 'true',
+        eventData: {
+            version: 'restv2.1'
+        },
+        events: args.events
+    };
+    let data = await connectApi.createConfiguration(args.accountId, { connectCustomConfiguration });
+    let { connectId, urlToPublishTo } = data;
+    return { connectId, urlToPublishTo };
+};
+
+/**
+ * Delete registered webhook.
+ * @param {Object} args
+ * @param {String} accessToken
+ * @param {String} connectId
+ * @return {Promise}
+ */
+const unregisterDocusignWebhook = async (args, accessToken, connectId) => {
+
+    let dsApiClient = new docusign.ApiClient();
+    dsApiClient.setBasePath(args.basePath + '/restapi/');
+    dsApiClient.addDefaultHeader('Authorization', 'Bearer ' + accessToken);
+    let connectApi = new docusign.ConnectApi(dsApiClient);
+    await connectApi.deleteConfiguration(args.accountId, connectId);
+};
+
+/**
+ * Normalizes multiselect field values to ensure consistent array format.
+ * This function handles various input formats (strings, arrays, comma-separated values)
+ * and converts them to a proper array format for API consumption.
+ *
+ * @param {string|Array|undefined} value - The value to normalize
+ * @returns {Array|undefined} - Normalized array or undefined if input is empty
+ */
+const normalizeMultiselect = (value) => {
+
+    if (!value) return undefined;
+
+    // If already an array, return as is
+    if (Array.isArray(value)) return value;
+
+    // If string, split by comma and trim whitespace
+    if (typeof value === 'string') {
+        const result = value.split(',').map(item => item.trim()).filter(item => item.length > 0);
         return result.length > 0 ? result : undefined;
     }
+
+    // For other types, convert to string first then split
+    const result = String(value).split(',').map(item => item.trim()).filter(item => item.length > 0);
+    return result.length > 0 ? result : undefined;
+};
+
+module.exports = {
+    getAccessToken,
+    refreshToken,
+    getEnvelope,
+    requestSignature,
+    registerDocusignWebhook,
+    unregisterDocusignWebhook,
+    normalizeMultiselect
 };
