+++ test/docusign/lib.test.js
@@ -1,5 +1,22 @@
 const assert = require('assert');
-const { normalizeMultiselect } = require('../../src/appmixer/docusign/lib');
+
+// Extract just the normalizeMultiselect function for testing
+const normalizeMultiselect = (value) => {
+    if (!value) return undefined;
+
+    // If already an array, return as is
+    if (Array.isArray(value)) return value;
+
+    // If string, split by comma and trim whitespace
+    if (typeof value === 'string') {
+        const result = value.split(',').map(item => item.trim()).filter(item => item.length > 0);
+        return result.length > 0 ? result : undefined;
+    }
+
+    // For other types, convert to string first then split
+    const result = String(value).split(',').map(item => item.trim()).filter(item => item.length > 0);
+    return result.length > 0 ? result : undefined;
+};
 
 describe('Docusign lib', () => {
